+++ CHANGELOG.md
@@ -29,12 +29,13 @@ This release has an [MSRV][] of 1.88.
 - The `any_true`, `all_true`, `any_false`, and `all_false` methods on mask types. ([#141][] by [@valadaptive][])
 - Documentation for most traits, vector types, and operations. ([#154][] by [@valadaptive][])
 - A "shift left by vector" operation, to go with the existing "shift right by vector". ([#155][] by [@valadaptive][])
+- "Precise" float-to-integer conversions, which saturate out-of-bounds results and convert NaN to 0 across all platforms. ([#167][] by [@valadaptive][])
 
 ### Fixed
 
 - Integer equality comparisons now function properly on x86. Previously, they performed "greater than" comparisons.
   ([#115][] by [@valadaptive][])
-- All float-to-integer and integer-to-float conversions are implemented properly on x86. They should now handle all values correctly, including NaN. ([#134][] by [@valadaptive][])
+- All float-to-integer and integer-to-float conversions are implemented properly on x86, including the precise versions. ([#134][] by [@valadaptive][])
 - The floating-point `min_precise` and `max_precise` operations now behave the same way on x86 and WebAssembly as they do on AArch64, returning the non-NaN operand if one operand is NaN and the other is not. Previously, they returned the second operand if either was NaN. ([#136][] by [@valadaptive][])
 
 ### Changed
@@ -154,6 +155,7 @@ No changelog was kept for this release.
 [#154]: https://github.com/linebender/fearless_simd/pull/154
 [#155]: https://github.com/linebender/fearless_simd/pull/155
 [#158]: https://github.com/linebender/fearless_simd/pull/158
+[#167]: https://github.com/linebender/fearless_simd/pull/167
 
 [Unreleased]: https://github.com/linebender/fearless_simd/compare/v0.3.0...HEAD
 [0.3.0]: https://github.com/linebender/fearless_simd/compare/v0.3.0...v0.2.0